+++ setup.cfg
@@ -7,7 +7,7 @@ tag = True
 search = version='{current_version}'
 replace = version='{new_version}'
 
-[bumpversion:file:ouroboros/__init__.py]
+[bumpversion:file:ouro/__init__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
