--- test-README.md
@@ -1,10 +1,0 @@
-# Ouroboros
-
-<p align="center">
-    <a href="" target="_blank">
-        <img
-            width="100%"
-            src="https://img.shields.io/badge/OUROBOROS-1e1e1e?style=for-the-badge&logo=data%3Aimage%2Fsvg%2Bxml%3Bbase64%2CPHN2ZyB2ZXJzaW9uPSIxLjIiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgdmlld0JveD0iMCAwIDEzNiAyNzciIHdpZHRoPSIxMzYiIGhlaWdodD0iMjc3Ij48ZGVmcz48bGluZWFyR3JhZGllbnQgaWQ9IlAiIGdyYWRpZW50VW5pdHM9InVzZXJTcGFjZU9uVXNlIi8%2BPGxpbmVhckdyYWRpZW50IGlkPSJnMSIgeDI9IjEiIGhyZWY9IiNQIiBncmFkaWVudFRyYW5zZm9ybT0ibWF0cml4KC05MS41OTgsNDEuNjM1LC02OC43MzEsLTE1MS4yMDksNzEuNzYsMTY3LjcxNikiPjxzdG9wIHN0b3AtY29sb3I9IiNmZmQ0M2IiLz48c3RvcCBvZmZzZXQ9IjEiIHN0b3AtY29sb3I9IiNmZmU4NzMiLz48L2xpbmVhckdyYWRpZW50PjxsaW5lYXJHcmFkaWVudCBpZD0iZzIiIHgyPSIxIiBocmVmPSIjUCIgZ3JhZGllbnRUcmFuc2Zvcm09Im1hdHJpeCgtODYuNDcyLDEyMi4zNDQsLTE1MS41MzYsLTEwNy4xMDUsMjY0LjQzNywxNDYuMzUyKSI%2BPHN0b3Agc3RvcC1jb2xvcj0iIzVhOWZkNCIvPjxzdG9wIG9mZnNldD0iMSIgc3RvcC1jb2xvcj0iIzMwNjk5OCIvPjwvbGluZWFyR3JhZGllbnQ%2BPC9kZWZzPjxzdHlsZT4uYXtmaWxsOnVybCgjZzEpfS5ie2ZpbGw6dXJsKCNnMil9PC9zdHlsZT48cGF0aCBmaWxsLXJ1bGU9ImV2ZW5vZGQiIGNsYXNzPSJhIiBkPSJtMzMuOCAyNzEuM2MtMjAuMy02LjctMzMuOC0xNC0zMy44LTMzLjN2LTE2Ni42aDI5LjZjMTguMyAwIDMzLjkgMTUuMSAzMy45IDMzLjN2NjYuNmMwIDIyLjIgMTkuMyA0MS42IDQyLjIgNDEuNmgyOS43djI1LjFjMCAxOS4zLTE0LjYgMjguNS0zMy44IDMzLjMtMjYuOCA2LjctNDYuOSA2LjktNjcuOCAwem0xMC43LTE2OWMwLTYuOS01LjctMTIuNS0xMi43LTEyLjUtNy4xIDAtMTIuOCA1LjYtMTIuOCAxMi41IDAgNi45IDUuNyAxMi41IDEyLjggMTIuNSA3IDAgMTIuNy01LjYgMTIuNy0xMi41eiIvPjxwYXRoIGZpbGwtcnVsZT0iZXZlbm9kZCIgY2xhc3M9ImIiIGQ9Im0xMDEuNSA1LjFjMjAuMyA2LjcgMzMuOSAxNCAzMy45IDMzLjN2MTY2LjZoLTI5LjdjLTE4LjIgMC0zMy44LTE1LjEtMzMuOC0zMy4zdi02Ni42YzAtMjIuMi0xOS4zLTQxLjYtNDIuMy00MS42aC0yOS42di0yNS4xYzAtMTkuNCAxNC41LTI4LjUgMzMuOC0zMy4zIDI2LjgtNi43IDQ2LjktNi45IDY3LjcgMHptMTQuOCAxNjljMC02LjktNS43LTEyLjUtMTIuNy0xMi41LTcgMC0xMi43IDUuNi0xMi43IDEyLjUgMCA2LjkgNS43IDEyLjUgMTIuNyAxMi41IDcgMCAxMi43LTUuNiAxMi43LTEyLjV6Ii8%2BPC9zdmc%2B"
-        />
-    </a>
-</p>
