+++ tox.ini
@@ -4,12 +4,12 @@ envlist = py38, flake8, black
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 ouroboros tests
+commands = flake8 ouro tests
 
 [testenv:black]
 basepython = python
 deps = black
-commands = black --check ouroboros tests
+commands = black --check ouro tests
 
 [testenv]
 setenv =
